SOLVED-PROBLEMS
Problem:
Write a program that does the following

‘A. Ask the user for values for four variables that store the following items of
data.
© A class name
© Test #1 score
© Test #2 score
© Test #3 score

Be sure to use appropriate prompts and variable names.

B. Compute the average of the three scores and store it in a variable.

C. Using the print function and format functions and your variables, generate
the following output, reporting back the input values and the average. Note
the formatting in the sample output below - make your program output
match the formatting. Be sure to use the format function and patterns
introduced in class. The actual values will depend on what is entered as the
input.

‘Scores Report For: Introduction to Computer Programming
TEST#1: 94
TEST#1: 76
TEST#1: 87

Average score: 85.7